修改后台每日总额

+++ src/com/liaoliao/profit/service/FenrunLogService.java
@@ -22,6 +22,6 @@ public interface FenrunLogService {
 
 	Integer countSignNum(Integer userId);
 
-	Integer todayTotal(Integer userId);
+	Double todayTotal(Integer userId);
 
 }
